% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_latency_efficiency.R
\name{add_latency_efficiency}
\alias{add_latency_efficiency}
\title{Adds sleep latency and efficiency to GGIR reports}
\usage{
add_latency_efficiency(outputfolder)
}
\arguments{
\item{outputfolder}{}
}
\value{
Does not retunr anything. Instead, sleep latency and efficiency will be added to the part 4 and part 5 reports from GGIR (both full and clean reports)
}
\description{
Adds sleep latency and efficiency to GGIR reports when they are not calculated (i.e., when sleeptypeWindow != "TimeInBed")
}
\details{
Caution! Sleep latency and efficiency depend in an accurate definition of the time in bed, which is not possible unless participants report the time to go to bed and get out of the bed in their sleep diaries.
}
