% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_naplog.R
\name{check_naplog}
\alias{check_naplog}
\title{Check the readibility of a nap log by the GGIR package}
\usage{
check_naplog(
  loglocation,
  coln1 = c(),
  colid = c(),
  nnights = c(),
  sleeplogidnum = TRUE,
  sleeplogsep = ",",
  desiredtz = "",
  meta.sleep.folder = c(),
  NapLowerLimit = 15/60,
  NapUpperLimit = 2,
  outputfile = c()
)
}
\arguments{
\item{loglocation}{Path to csv file containing the sleeplog (see details)}

\item{coln1}{Column number in the sleep log spreadsheet where the onset of the first night starts}

\item{colid}{Column number in the sleep log spreadsheet in which the participant ID code is stored (default = 1)}

\item{nnights}{Number of nights for which sleep log information should be available. It assumes that this is constant within a study. If sleep log information is missing for certain nights then leave these blank}

\item{sleeplogidnum}{Should the participant identifier as stored in the sleeplog be interpretted as a number (TRUE=default) or a character (FALSE)?}

\item{sleeplogsep}{Value used as sep argument for reading sleeplog csv file, usually "," or ";".}

\item{desiredtz}{Character (default = "", i.e., system timezone). Timezone in which device was configured and experiments took place. If experiments took place in a different timezone, then use this argument for the timezone in which the experiments took place and argument configtz to specify where the device was configured. See also https://en.wikipedia.org/wiki/Zone.tab}

\item{meta.sleep.folder}{Path to part3 milestone data, only specify if sleeplog is in advanced format.}

\item{NapLowerLimit}{Limit to define a extremely short nap in hours}

\item{NapUpperLimit}{Limit to define a extremely long nap in hours}

\item{outputfile}{If defined, path to xlsx file to store the output}
}
\value{
Save excel file with extremely short and long nights, daysleeper nights, and missing nigths from sleeplog.
}
\description{
Function to try to read a nap log from an advanced sleeplog (csv file) by the GGIR package
}
