% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sleeplog.R
\name{check_sleeplog}
\alias{check_sleeplog}
\title{Check the readibility of a sleeplog by the GGIR package}
\usage{
check_sleeplog(
  loglocation,
  coln1 = c(),
  colid = c(),
  desiredtz = "",
  meta.sleep.folder = c(),
  SPTlowerLimit = 4,
  SPTupperLimit = 10,
  advanced_sleeplog = FALSE,
  outputfile = c()
)
}
\arguments{
\item{loglocation}{Path to csv file containing the sleeplog (see details)}

\item{coln1}{Column number in the sleep log spreadsheet where the onset of the first night starts}

\item{colid}{Column number in the sleep log spreadsheet in which the participant ID code is stored (default = 1)}

\item{desiredtz}{Character (default = "", i.e., system timezone). Timezone in which device was configured and experiments took place. If experiments took place in a different timezone, then use this argument for the timezone in which the experiments took place and argument configtz to specify where the device was configured. See also https://en.wikipedia.org/wiki/Zone.tab}

\item{meta.sleep.folder}{Path to part3 milestone data, only specify if sleeplog is in advanced format.}

\item{SPTlowerLimit}{Limit to define a extremely short night in hours}

\item{SPTupperLimit}{Limit to define a extremely long night in hours}

\item{advanced_sleeplog}{Logical, if sleeplog is formatted as advanced or basic GGIR format}

\item{outputfile}{If defined, path to xlsx file to store the output}
}
\value{
Save excel file with extremely short and long nights, daysleeper nights, and missing nigths from sleeplog.
}
\description{
Function to try to read a sleeplog (csv file) by the GGIR package
}
