% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_id_dates_sleeplog.R
\name{fix_id_dates_sleeplog}
\alias{fix_id_dates_sleeplog}
\title{Check that ID and dates are logical and fix them if needed}
\usage{
fix_id_dates_sleeplog(
  loglocation,
  colid,
  meta.sleep.folder,
  advanced_sleeplog = TRUE,
  dateformat = "\%d/\%m/\%Y",
  outputxlsx,
  outputGGIR
)
}
\arguments{
\item{loglocation}{Path to csv file containing the sleeplog (see details)}

\item{colid}{Column number in the sleep log spreadsheet in which the participant ID code is stored (default = 1)}

\item{meta.sleep.folder}{Path to part3 milestone data, only specify if sleeplog is in advanced format.}

\item{advanced_sleeplog}{Logical, if sleeplog is formatted as advanced or basic GGIR format}

\item{dateformat}{Format of date in sleeplog as in \link{strptime}}

\item{outputxlsx}{If provided, path to xlsx file to save the revised sleeplog with tracked changes}

\item{outputGGIR}{Path to csv file for sleeplog to be read by GGIR}
}
\value{
Save excel and csv file with new sleeplog to be read by GGIR
}
\description{
Check that ID and dates are logical and fix them if needed
}
