% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGIR_automatic_vs_sleeplog.R
\name{GGIR_automatic_vs_sleeplog}
\alias{GGIR_automatic_vs_sleeplog}
\title{Title}
\usage{
GGIR_automatic_vs_sleeplog(
  outputfolder,
  f0 = 1,
  f1 = 0,
  loglocation,
  coln1 = c(),
  colid = c(),
  nnights = c(),
  sleeplogidnum = TRUE,
  sleeplogsep = ",",
  desiredtz = "",
  sleepwindowType = "SPT",
  criterror = 2,
  outputfile = c(),
  outputplotname = c()
)
}
\arguments{
\item{sleepwindowType}{}
}
\description{
Title
}
