% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACCutils-package.R
\docType{package}
\name{package1}
\alias{package1}
\title{ACCutils Package}
\description{
Functions and utilities that I use frequently to process and quality check the accelerometer data collected with different brands.
}
\author{
Jairo Hidalgo Migueles \email{jairo.hidalgo.migueles@gmail.com}
}
