% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postGGIR_revise_sleeplog.R
\name{postGGIR_revise_sleeplog}
\alias{postGGIR_revise_sleeplog}
\title{Title}
\usage{
postGGIR_revise_sleeplog(
  outputfolder,
  criterror = 1,
  outputfile,
  outputplotname
)
}
\arguments{
\item{outputplotname}{}
}
\description{
Title
}
