% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postGGIR_revise_SPT.R
\name{postGGIR_revise_SPT}
\alias{postGGIR_revise_SPT}
\title{Title}
\usage{
postGGIR_revise_SPT(
  outputfolder,
  desiredtz,
  criterror = 4,
  SPTlowerLimit = 4,
  SPTupperLimit = 10,
  excludefirstlast = FALSE,
  method = c("all", "last")[1],
  outputfile,
  outputplotname
)
}
\arguments{
\item{method}{}
}
\description{
Title
}
