% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_conflicting_nights.R
\name{visualize_conflicting_nights}
\alias{visualize_conflicting_nights}
\title{Title}
\usage{
visualize_conflicting_nights(
  outputfolder,
  NS,
  NSconflicts,
  desiredtz = "Europe/Madrid",
  onset_0 = c("sleeponset_ts", "sleeponset_ts_auto")[2],
  onset_1 = c("guider_onset_ts", "guider_inbedStart_ts", "sleeponset_ts_sleeplog")[3],
  wake_0 = c("wakeup_ts", "wakeup_ts_auto")[2],
  wake_1 = c("guider_wakeup_ts", "guider_inbedEnd_ts", "wakeup_ts_sleeplog")[3],
  outputplotname
)
}
\arguments{
\item{wake_1}{}
}
\description{
Title
}
