% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_conflicting_nights.R
\name{visualize_conflicting_nights}
\alias{visualize_conflicting_nights}
\title{Title}
\usage{
visualize_conflicting_nights(
  outputfolder,
  NSconflicts,
  desiredtz = "Europe/Madrid",
  outputplotname
)
}
\arguments{
\item{outputplotname}{}
}
\description{
Title
}
